+++ packages/components/src/scrollbar/info.ts
@@ -1,0 +1,5 @@
+import { defineDocInfo } from '@/.vitepress/types/info';
+
+export default defineDocInfo({
+  name: '滚动条',
+});