+++ packages/components/src/scrollbar/src/types.ts
@@ -1,0 +1,8 @@
+import type { PartialOptions } from 'overlayscrollbars';
+
+export interface ScrollbarProps {
+  /** 滚动条选项 */
+  options?: PartialOptions;
+  /** 滚动条容器样式类 */
+  contentClass?: string;
+}